pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "ganzoo"
version = "0.1.0"
description = "Collection of TF2 implementation of Generative Adversarial Networks."
authors = ["LIU Jialin <jialinliu_ml@163.com>"]
readme = "README.md"
license = "LICENSE"
packages = [
    {include = "ganzoo"},
]
repository = "https://github.com/limtete/GANZoo"
keywords = ["Generative Adversarial Networks", "Tensorflow 2.0", "Keras"]

[tool.poetry.dependencies]
python = "~3.9.0"
tensorflow = "2.9.0"
tensorflow-datasets = "4.6.0"
absl-py = "1.2.0"
pytest = "^7.1.3"

[tool.pytest.ini_options]
addopts = [
    "--import-mode=importlib",
]
testpaths = [
    "tests",
]
filterwarnings = [
    "ignore::DeprecationWarning",
]